Moved backup system out of main into lib.
