+++ main.js
@@ -20,6 +20,8 @@ const settings = require('./lib/interactive/settings-access').settings;
 
 const dataAccess = require('./lib/data-access.js');
 
+const backupManager = require("./lib/backupManager");
+
 var ncp = require('ncp').ncp;
 ncp.limit = 16;
 
@@ -254,6 +256,7 @@ function createWindow () {
   app.on('will-quit', () => {
     // Unregister all shortcuts.
     mixerConnect.shortcutUnregister();
+    if(settings.backupOnExit()) backupManager.startBackupManager();
   });
   
   // Run Updater
@@ -334,62 +337,9 @@ function createWindow () {
     shell.showItemInFolder(backupFolder);
   });
 
-
-  // Backup Routine
-  // 
-  ipcMain.on('startBackup', function(event, onUser = false){
-    var archiver = require('archiver');
-    var backupKeepAll = settings.backupKeepAll();
-    var backupOnExit = settings.backupOnExit();
-    var backupBeforeUpdates = settings.backupBeforeUpdates();
-    var timestamp = Date.now();
-    var fileExtension = 'zip';
-
-    if(backupKeepAll){
-      var filename = 'backup-' + timestamp +'.' + fileExtension;
-    }else{
-      console.log("Overwriting the same file each time...");
-      var filename = 'backup.'+ fileExtension;
-    }
-    var backupPath = path.resolve(dataAccess.getUserDataPath() + path.sep + "backups");
-    var folderPath = path.resolve(dataAccess.getUserDataPath() + path.sep + "user-settings");
-    var output = fs.createWriteStream(backupPath + path.sep + filename);
-    var archive = archiver(fileExtension, {
-      zlib: { level: 9 } // Sets the compression level.
-    });
-    
-    // listen for all archive data to be written
-    output.on('close', function() {
-      filesize = archive.pointer() / 1000;
-      if(onUser) renderWindow.webContents.send('info', "The backup has been finished, you can close this message safely. (Backup size: "+ filesize +" bytes)");
-    });
-
-    archive.on('warning', function(err) {
-      if (err.code === 'ENOENT') {
-          // log warning
-          if(onUser) renderWindow.webContents.send('error', "There was an error initiating the backup function.");
-          renderWindow.webContents.send('error', err);
-      } else {
-          // throw error
-          if(onUser) renderWindow.webContents.send('error', "Something bad happened, please check your logs.");
-          renderWindow.webContents.send('error', err);
-          throw err;
-      }
-    });
-
-    // good practice to catch this error explicitly
-    archive.on('error', function(err) {
-      throw err;
-    });
-
-    // pipe archive data to the file
-    archive.pipe(output);
-    
-    // Add directory to package
-    archive.directory(folderPath,'user-settings');
-    
-    // finalize the archive (ie we are done appending files but streams have to finish yet)
-    archive.finalize();
+  //ipcMain.on('startBackup');
+  ipcMain.on('startBackup', function(event, manualActivation = false){
+    backupManager.startBackupManager(manualActivation);
   });
 
 // In this file you can include the rest of your app's specific main process
